Out with the new, in with the newer. + SearchEngine, templated on node storage, housing A* search templated on goal, cost & heuristic functions. + Cartographer & InfluenceMap introduced (neither very functional yet) + PathFinder renamed PathManager (for now anyway, will be RoutePlanner soon I think :-) + DebugRenderer added, has templated functions to draw alternate textures for cells, or an 'overlay' on top of the cells. + Moved as much debugging texture/overlay related stuff into the DebugRenderer as possible. + changed to DEBUG_THIS_THAT style conditional compiling defines. + added rudimentary (explicit) call counter to Profiler + sprinkled Profiling macros all over the place

+++ source/glest_game/main/program.cpp
@@ -92,10 +92,10 @@ Program *Program::singleton = NULL;
 // ===================== PUBLIC ========================
 
 Program::Program(Config &config, int argc, char** argv) :
-		renderTimer(config.getRenderFpsMax(), 1),
+		renderTimer((float)config.getRenderFpsMax(), 1),
 		tickTimer(1.f, maxTimes, -1),
-		updateTimer(config.getGsWorldUpdateFps(), maxTimes, 2),
-		updateCameraTimer(GameConstants::cameraFps, maxTimes, 10),
+		updateTimer((float)config.getGsWorldUpdateFps(), maxTimes, 2),
+		updateCameraTimer((float)GameConstants::cameraFps, maxTimes, 10),
 		programState(NULL),
 		crashed(false),
 		keymap(getInput(), "keymap.ini") {
